# lib/api-zod/src/generated/types/bulkProductUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkProductUpdateCategory } from './bulkProductUpdateCategory';
import type { BulkProductUpdateGender } from './bulkProductUpdateGender';

export interface BulkProductUpdate {
  /** @minItems 1 */
  productIds: string[];
  /** @nullable */
  gender?: BulkProductUpdateGender;
  /** @nullable */
  category?: BulkProductUpdateCategory;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  reviewed?: boolean | null;
  /** @nullable */
  isPublished?: boolean | null;
}