# lib/api-zod/src/generated/types/catalogueResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { Collection } from './collection';
import type { Product } from './product';

export interface CatalogueResult {
  collection: Collection;
  products: Product[];
  availableBrands: string[];
}