# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
  isPublished: boolean;
  /** @nullable */
  publishedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}