# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  endDate?: Date | null;
}