# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { Collection } from './collection';
import type { DashboardSummaryCategories } from './dashboardSummaryCategories';

export interface DashboardSummary {
  collection: Collection;
  totalUploaded: number;
  men: number;
  women: number;
  unknown: number;
  needsReview: number;
  published: number;
  categories: DashboardSummaryCategories;
}