# lib/api-zod/src/generated/types/pdfRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfRequest {
  title: string;
  /** @minItems 1 */
  productIds: string[];
}