# lib/api-zod/src/generated/types/pdfResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PdfResultStatus } from './pdfResultStatus';

export interface PdfResult {
  status: PdfResultStatus;
  /** @nullable */
  downloadUrl?: string | null;
  title: string;
}