# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';
import type { ProductGender } from './productGender';
import type { ProductImage } from './productImage';

export interface Product {
  id: string;
  collectionId: string;
  gender: ProductGender;
  category: ProductCategory;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  aiGender?: string | null;
  /** @nullable */
  aiCategory?: string | null;
  /** @nullable */
  aiBrand?: string | null;
  /** @nullable */
  aiConfidence?: number | null;
  reviewed: boolean;
  isActive: boolean;
  isPublished: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
  images: ProductImage[];
}