# lib/api-zod/src/generated/types/productImage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductImage {
  id: string;
  imagePath: string;
  isPrimary: boolean;
  sortOrder: number;
}