# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCategory } from './productInputCategory';
import type { ProductInputGender } from './productInputGender';

export interface ProductInput {
  collectionId: string;
  gender: ProductInputGender;
  category: ProductInputCategory;
  /** @nullable */
  brand?: string | null;
  imagePath: string;
  /** @nullable */
  imageUrl?: string | null;
}