# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  whatsappNumber: string;
  businessName: string;
  tagline: string;
}